feat(user): 탈퇴시 비밀번호 검사 추가

+++ src/main/java/com/example/usermanagement/service/UserService.java
@@ -90,6 +90,10 @@ public class UserService {
             return new UserSignupPayload(0L);
         }
 
+        if (!passwordEncoder.matches(userDto.getPassword(), optUser.get().getPassword())) {
+            return new UserSignupPayload(-1L);
+        }
+
         userRepository.delete(optUser.get());
 
         return new UserSignupPayload(optUser.get().getUserId());
